Refactored SkyDrive to OneDrive
