+++ KeePass/Sources/OneDrive/List.xaml
@@ -1,5 +1,5 @@
 ﻿<keepass:KeePassPage 
-    x:Class="KeePass.Sources.SkyDrive.List"
+    x:Class="KeePass.Sources.OneDrive.List"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:keepass="clr-namespace:KeePass.Utils" xmlns:uc="clr-namespace:KeePass.Controls"
@@ -16,7 +16,7 @@
         <StackPanel x:Name="TitlePanel" Grid.Row="0" Margin="12,17,0,28">
             <TextBlock x:Name="ApplicationTitle" Text="{StaticResource AppName}"
                        Style="{StaticResource PhoneTextNormalStyle}"/>
-            <TextBlock x:Name="PageTitle" Text="SkyDrive" Margin="9,-7,0,0"
+            <TextBlock x:Name="PageTitle" Text="OneDrive" Margin="9,-7,0,0"
                        Style="{StaticResource PhoneTextTitle1Style}"/>
         </StackPanel>
