+++ KeePass/Sources/OneDrive/LiveAuth.xaml
@@ -1,5 +1,5 @@
 ﻿<keepass:KeePassPage 
-    x:Class="KeePass.Sources.SkyDrive.LiveAuth"
+    x:Class="KeePass.Sources.OneDrive.LiveAuth"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:phone="clr-namespace:Microsoft.Phone.Controls;assembly=Microsoft.Phone"
